// src/ui/generated/graphql.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string | number;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Invoice = {
  __typename?: 'Invoice';
  id: Scalars['ID'];
  createdAt?: Maybe<Scalars['DateTime']>;
  orderCode: Scalars['String'];
  orderId: Scalars['String'];
  customerEmail: Scalars['String'];
  invoiceNumber: Scalars['Int'];
  downloadUrl: Scalars['String'];
};

export type InvoiceConfig = {
  __typename?: 'InvoiceConfig';
  id: Scalars['ID'];
  enabled: Scalars['Boolean'];
  templateString?: Maybe<Scalars['String']>;
};

export type InvoiceConfigInput = {
  enabled: Scalars['Boolean'];
  templateString?: Maybe<Scalars['String']>;
};

export type InvoiceList = {
  __typename?: 'InvoiceList';
  items: Array<Invoice>;
  totalItems: Scalars['Int'];
};

export type InvoicesListInput = {
  page: Scalars['Int'];
  itemsPerPage: Scalars['Int'];
};
// ...
export type MyInvoicesInput = {
  customerEmail: Scalars['String'];
  list?: Maybe<InvoicesListInput>;
};